made first version of parser, didn`t check for working with variables

+++ Task_1_1_3/src/main/java/ru/nsu/chuvashov/expressionparser/Number.java
@@ -14,7 +14,7 @@ public class Number extends Expression {
      *
      * @param number - number that we use.
      */
-    public Number(int number) {
+    public Number(double number) {
         this.value = number;
     }
